Fixes logic in dictionary describe
The `dictionary describe --verbose` command makes additional calls to
the fastly api's GetDictionaryInfo and ListDictionaryItems endpoints.
The call to GetDictionaryInfo was using the serviceID value in place of
the dictionaryID, which would have resulted in a failed call at best and
values from the wrong dictionary at worst.

I added enforcement in the mock call to that endpoint to verify that the
dictionaryID in the input is the same as the dictionaryID in the output
from the original call to GetDictionary.