Logs Tail: Give the user better feedback when --from flag errors
This fixes two things:
- When an unknown status code is returned we previously return from
the function without and exit. This would just leave us in and endless
loop doing nothing for the user.
- When a user specifies a --from time that is too far in the past or
future, we should let them know that so it doesn't just mysteriously
give them a 404 warning and exit.

Previously it would respond like the following and just hang:
```
WARNING: non-200 resp 404
```

With this change it looks errors and exits:
```
ERROR: specified from time 1622362928 not found, either too far in the past or future
```