validate manifest_version is supported...
This is actually a complex change because we need to do a bunch of work
that is done by the custom Version type when calling toml.Unmarshal().

The reason we have to duplicate that work is because we can find
ourselves in a position where toml.Unmarshal will error if given a toml
file with a structure not supported by the CLI code, and so in order to
provide a good remediation to the user we need to first identify the
manifest_version separate from a full unmarshal of the data.