Ignore possible `rustup` output in `rustc --version`. Closes #452
If the `rustc` executable that the CLI invokes is managed by `rustup`,
*and* the Rust toolchain that `rustup` considers active for the PWD is
not installed, then the first several lines of output are from `rustup`
installing the toolchain, and only after it finishes does `rustup`
execute the real `rustc --version` and print its output.

This patch collects and throws away all but the last output lines from
`rustc --version`, since the last line is *always* the version string.

It has been tested to work by using `fastly compute init`, changing the
`rust-toolchain` file to a named version not installed, then running
`fastly compute serve`. This caused an installation, then was able to
correctly parse the output of `rustc --version` and continue.