Makefile: when building binary, depend on .go files
If you made a change to a source file and ran `make fastly`, it'd say
that fastly was up-to-date because the makefile target did not depend on
its input files.  Fix that by creating a list of all the Go source files
and having the makefile target depend on it.