fix(errors/log): support allowing errors to Sentry (#909)
* fix(errors/log): support allowing errors to Sentry

The behaviour is now reversed. We no longer default to sending all errors to
Sentry. Instead, for each error report, we must opt-in to sending the error to
Sentry. This is because the majority of errors instrumented are Fastly API
errors and these are not actionable as far as the Fastly CLI is concerned.

* fix: AllowInstrumentation