Send compute events to the API
In order to be able to update the UI to reflect the full process of
starting a compute service, we need to know when certain events have
occurred. Rather than polling, this creates a new internal package that
uses undocumented to make pings back to the API for different events.
The important events at the moment are compute init when using an
existing service ID and activating a service version during deploy.