fix(compute/rust) handling of 'cargo version' output (#1197)
'cargo version' emits the version string to stdout, so the code should
capture only stdout, and parse it for checking the
ToolchainConstraint. Capturing stderr and including it in the parsing
process can result in bogus version numbers being found and reported
(as is happening right now when Rust 1.78 is used in a project
directory where .cargo/config exists).