feat(compute/rust) Handle Cargo config filename for Rust >=1.78.0
The preferred Cargo configuration filename has been 'config.toml'
since Rust 1.39.0 was released, but 'config' was supported for
backwards compatibility (although deprecated).

In Rust 1.78.0, the backwards compatibility was removed, so any file
with the old name will be ignored. Since Compute package builds
require the content in that file, ignoring it results in a build
failure.

This patch adds checking for this situation. If the old file name is
found, and the Rust version is <1.78.0, a warning is issued to let the
user know that they should rename the file before upgrading to 1.78.0
or later.  If the old file name is found, and the Rust version is
>=1.78.0, an error is emitted and the build process stopped, as it
will not succeed.