Adds --cert-path argument to tls-custom certificate update command to pass in a path to a certificate file (#1214)
Previously users were required to pass in the full certificate contents via the --cert-blob argument, which is cumbersome and error-prone. This allows users to specify a path to a certificate making this command easier to use.

The new argument --cert-path is mutually exclusive with --cert-blob and the command will return an error if both are specified.