refactor: Simplify declaration and execution of test scenarios
Adds an 'Arg' field to the TestScenario structure, to hold a simple
string of arguments instead of a slice of individual arguments.

Adds a 'RunScenarios' function to the testutil package which is able
to execute all of the tests in a slice of TestScenario structs. This
function accepts an additional slice (of strings) which indicate the
command name(s) to be prepended to the arguments supplied in each
scenario, so that they do not need to be repeated in the scenario
declarations. Finally, it splits the 'Arg' field into individual
arguments so that the scenario definition can supply a simple string.

Converts serviceversion_test.go to use the new features.

Converts .tmpl/test.go to use the new features, and also fixes some
problems with the test scenarios declared there.

The remaining tests will be converted to the new features once this PR
has been merged.