deps: Add 'optional' package
This adds the 'optional' package from
https://github.com/leighmcculloch/go-optional.

The package has been copied into a 'thirdparty' directory because it
is not directly installable from the GitHub repository, its canonical
location is the developer's personal website which is not ideal from a
reproducibility/safety point of view. It was not placed into 'vendor'
because 'go mod' can manipulate that directory if it exists.