fix(all commands): --service-name flag should have priority.
When the --service-name flag was added it was given the lowest
priority for obtaining a service ID. If a user supplied this flag but
also supplied --service-id, or had a fastly.toml with a service ID in
it, or had the FASTLY_SERVICE_ID environment variable set, the
--service-name flag would be ignored. This resulted in a user
accidentally deleting a service because the flag was ignored.

This commit changes the priority so that --service-name and
--service-id have equal (top) priority, and also ensures that
supplying both will result in an error. It also adds a unit test for
the argparser.ServiceID() function which implements this logic.

Closes #1261.