fix(service-version): Allow 'locked' services to be activated. (#1245)
* deps: Add 'optional' package.

* fix(service-version): Allow 'locked' services to be activated.

This PR changes the ServiceDetailsOpts structure (used by the
ServiceDetails function) to permit commands to specify individual
serviceversion states as part of the filtering mechanism. Nearly all
commands allow both 'active' and 'locked' serviceversions, and
previously 'service-version activate' blocked both, but it should
allow 'locked' while still blocking 'active'.

Additionally, the error message emitted when the specified
serviceversion is not allowed by the filter now indicates whether it
was the 'active' or 'locked' status which caused the error.

Finally, many more tests were added to ensure that the proper error
messages are emitted for 'active' and 'locked' serviceversions in many
of the commands which specify filters.