fix(profile/token): 'profile token' command must check the validity of the stored token.
The 'fastly profile token' command did not check the validity
(expiration) of the stored token, which meant that it would emit an
invalid token if the stored session token (and refresh token) had
expired. This PR changes the behavior so that the validity of the
token is checked as it is for all commands which actually make use of
the token.