feat(profile/token): Allow user to specify how long token must be valid. (#1340)
* feat(profile/token): Allow user to specify how long token must be valid.

The default token validity period for 'fastly profile token' is 5
minutes, but users may want to ensure that the token is valid for a
longer period (or only a shorter period). This PR adds 'fastly profile
token --ttl=<period>', allowing the user to specify the validity
period in seconds, minutes, or hours (or some combination).

* Restore support for non-OIDC tokens in `fastly profile token`.

Profiles with non-OIDC tokens (tokens generated using the traditional
Fastly token-creation methods) do not contain expiration information,
so the validity-checking in the `fastly profile token` command should
not attempt to validate them.

* Emit proper suggested command when an OIDC token needs to be refreshed.

The `fastly profile update` command is not the best suggestion in this
situation; the user should directly invoke `fastly sso` instead.