feat(compute/build): Support 'upper bound' constraints on Rust versions. (#1350)
* feat(compute/build): Support 'upper bound' constraints on Rust versions.

Changes planned in the upcoming Rust 1.84 release may not be
compatible with the Fastly Compute platform (until further analysis is
done, which may result in the need for platform changes). To ensure
that customers do not build Compute packages which do not operate
correctly in the platform, this PR adds a toolchain version constraint
"upper bound" in addition to the existing "lower bound", only for
Rust.

The upper bound is currently set to "<1.84.0", so any versions
currently available, plus all releases of Rust 1.83.x, will be
permitted. Because this is a hard requirement of the platform, the
check has been changed to report an error (not a warning), if the
version of Rust in use does not meet the requirements. While this is
potentially a breaking change (since the 'fastly compute build'
process will stop, instead of continuing), it can only affect users
with Rust versions older than 1.56.1, which are not actually usable at
this time anyway.

In addition, this PR changes the messages generated for both Go and
Rust version-constraint failures to include the details of how the
check failed.

* Improve error message for too-high Rust version.

Use a RemediationError to give the user a link to a documentation page
where they can get more information about supported versions.