fix(sso): Ensure that only one authentication cycle is started. (#1355)
The 'sso' command, when run against a profile with an expired refresh
token, would cause two authentication cycles to be started: one by the
CLI core before the 'sso' command itself was executed, and then
another by the 'sso' command. The second cycle would fail and result
in an 'extra' browser tab/window with an error message of "ERROR: no
authorization code returned". While the authentication was successful
and the new refresh/access/session tokens were stored, this confused
the user of the CLI.