Additional checking for wasm32-wasi output from Rust projects.
In version 11.0.0 a test was added to ensure that the CLI's global
configuration file was not configured for 'wasm32-wasi' output from
Rust projects. That was insufficient to avoid build errors in existing
projects configured for that target, so this commit adds an additional
check. Specifically, if the WASM binary is not found in the expected
location after a successful build, a check is made to see if the
binary is found in the 'old' location. If it is found there, an error
is emitted which includes a link to the instructions for how to
reconfigure the project.