Increase max config store value length
The maximum allowed length for a config store entry's value is now
configurable. The default is still 8k characters, but may be larger.
This increases the limit in the CLI to 64k. If a user attempts to create
an entry larger than their allowed limit, the API will return an error.