feat: Add experimental "enable Pushpin" mode (#1509)
This PR adds code to the CLI to enable `fastly compute serve` with
"enable Pushpin", a new mode that locates and runs Pushpin before
starting Viceroy to run guest code.

This is a companion PR to https://github.com/fastly/Viceroy/pull/497,
and will only function after that feature has landed.

## How to enable it:

It can be invoked in one of two ways:

1. set a new flag `--experimental-enable-pushpin`
2. or, set a new flag in `fastly.toml`:

```toml
[local_server.pushpin]
enable = true
```

## What it does:

* `fastly compute serve` starts Pushpin before starting Viceroy, and
detects its startup. It also sets up to shut it down when the CLI
process ends.
* `pushpin` must be started at this time, because routes are set based
on the backends defined in the `fastly.toml` manifest file.
* "kill process" is done differently between Windows and Unix, so build
flags are used to include the right implementation of `killProcess`.

* Pushpin routes are set up based on the `fastly.toml` manifest file.
    * A route is set up per backend.
* The name of each backend is set as an `id=` condition. For example, a
backend named `origin` will set `id=origin`. Viceroy handles Fanout
handoff by forwarding a request for this backend to Pushpin and set the
header `Pushpin-Route: origin`, allowing Pushpin to match the route.
* If the backend definition's URL has a path segment, then it is set as
a `replace_beg` configuration on the route. For example, if a backend is
configured for `http://localhost:3000/realtime`, then this is set as
`replace_beg=realtime`. Pushpin will forward a request for `/api/user`
to this backend as `http://localhost:3000/realtime/api/user`.
* `over_http` is set for every route target, as to always enable
WebSocket-over-HTTP.
    * If the backend URL is HTTPS, then `ssl` is added.
* If the backend definition sets `override_host`, then it will be set as
the `host` value of the route target.

* When starting Viceroy, `--local-pushpin-proxy-port=<port>` is appended
as a command-line argument, which starts Viceroy in "enable Pushpin"
mode.

* Pushpin proxy runs by default on port `7677`, but this can be
overridden:
* using `proxy_port=<port>` under `local_server.pushpin` section of the
`fastly.toml` file, or
   * using `--pushpin-proxy-port=<port>`

* Pushpin's publishing endpoint runs by default on port `5561`, but this
can be overridden:
* using `publish_port=<port>` under `local_server.pushpin` section of
the `fastly.toml` file, or
   * using `--pushpin-publish-port=<port>`

* The `pushpin` binary is searched on the system path, but can be
specified:
* using `pushpin_path=/path/to/pushpin` under the `local_server.pushpin`
section of the `fastly.toml` file, or
   * using `--pushpin-path=/path/to/pushpin`.

* The CLI writes the config and routes to temporary files in the OS
temporary directory. This is cleaned up as part of shutdown.

* Pushpin runner places some runtime files into a temporary directory
specified in pushpin.conf. We set this to an OS temporary directory, so
that multiple instances can run if necessary. This is cleaned up as part
of shutdown.

* Pushpin runner places logs into a directory specified in pushpin.conf.
This is placed in the ./pushpin-logs directory under the current project

All Submissions:

* [x] Have you followed the guidelines in our Contributing document?
* [x] Have you checked to ensure there aren't other open [Pull
Requests](https://github.com/fastly/cli/pulls) for the same
update/change?

### New Feature Submissions:

* [x] Does your submission pass tests?

### Changes to Core Features:

* [x] Have you added an explanation of what your changes do and why
you'd like us to include them?
* [ ] Have you written new tests for your core changes, as applicable?
* [x] Have you successfully run tests with your changes locally?

### User Impact

* [x] What is the user impact of this change?

### Are there any considerations that need to be addressed for release?

The new feature is opt-in using the `--experimental-enable-pushpin`
flag, and does nothing if the new flag is not set.