Upgrade to go-fastly version 11.
go-fastly version 11 requires a context.Context object to be passed as
the first argument in all of the API functions it exports. This commit
modifies all of the code in the CLI to pass context.TODO() in that
case, since the CLI does not use context objects for any other
purpose.

The CLI also has an internal 'interface' which maps to the complete
set of API functions in the 'fastly' namespace exported by go-fastly,
and hundreds of 'mock' functions used in the CLI test suite. Those
have also been updated, and along the way many inconsistencies in
those function signatures were corrected.

go-fastly version 11 removed a deprecated set of product enablement
API functions, so the 'fastly products' command has been updated to
use the new API functions. The tests for the `fastly products` command
no longer mock the API calls as they no longer exist, and the new
'fastly product' command which will appear in the next release will
provide full test coverage of the CLI's use of the new product
enablement API functions.

There are no functional changes in this commit, and the entire test
suite still passes.

At least 80% of the changes in this commit were produced using
[comby](https://comby.dev/).