Ensure [local_server.pushpin] section persists during fastly.toml manifest file update (#1535)
### Change summary

This PR updates `MarshalTOML()` for `manifest/File` such that the
`[local_server.pushpin]` section and its fields persist when the file is
written.

* The `pushpin` section as well as its individual fields are now
optional, such that it will be able to persist the file back as it was
before starting.
* `[local_server.pushpin]` has been added to
`fastly-viceroy-update.toml`, which is used by the test
`TestManifestPersistsLocalServerSection`, ensuring that the section is
written out and reads back in to the same value.

This fixes #1532.

* [x] Have you followed the guidelines in our Contributing document?
* [x] Have you checked to ensure there aren't other open [Pull
Requests](https://github.com/fastly/cli/pulls) for the same
update/change?

### Changes to Core Features:

* [x] Have you successfully run tests with your changes locally?

### User Impact

* [x] What is the user impact of this change?

Users can now safely run CLI commands that modify fastly.toml without
losing their `[local_server.pushpin]` section.

### Are there any considerations that need to be addressed for release?

Since this overwrite happens right after initializing an app from the
starter kit, this would be nice to get out asap.
Would be great if this can be released as a quick patch release such as
11.5.1.

---------

Co-authored-by: Richard Carillo <77027245+rcaril@users.noreply.github.com>