Enable support for pre-release Rust versions.
For Rust, set the `IncludePrerelease` field of the semver [`Constraints`]
to true. This makes it possible to set `channel = "nightly"`, without
getting errors like this:

```
ERROR: the Rust version requirement was not satisfied: '1.94.0-nightly is a prerelease version and the constraint is only looking for release versions'.
```

Most users won't have any reason to use `channel = "nightly"`, however it
is useful for users that specifically wish to test `channel = "nightly",
such as to help find problems with future Rust versions before they become
stable versions.

[`Constraints`]: https://pkg.go.dev/github.com/Masterminds/semver/v3#Constraints