Enable support for pre-release Rust versions. (#1604)
### Change summary

For Rust, set the `IncludePrerelease` field of the semver
[`Constraints`] to true. This makes it possible to set `channel =
"nightly"`, without getting errors like this:

```
ERROR: the Rust version requirement was not satisfied: '1.94.0-nightly is a prerelease version and the constraint is only looking for release versions'.
```

Most users won't have any reason to use `channel = "nightly"`, however
it is useful for users that specifically wish to test `channel =
"nightly", such as to help find problems with future Rust versions
before they become stable versions.

[`Constraints`]:
https://pkg.go.dev/github.com/Masterminds/semver/v3#Constraints

 All Submissions:

* [x] Have you followed the guidelines in our Contributing document?
* [x] Have you checked to ensure there aren't other open [Pull
Requests](https://github.com/fastly/cli/pulls) for the same
update/change?

### New Feature Submissions:

* [x] Does your submission pass tests?

### Changes to Core Features:

* [ ] Have you written new tests for your core changes, as applicable?
* [x] Have you successfully run tests with your changes locally?

I didn't write any tests because I didn't find any test harness in the
tree suitable for writing tests for a change of this nature, and also,
it's a minor change that only impacts users that opt into it by choosing
prerelease Rust versions, which will include me, for the purpose of
testing.

### User Impact

The only impact is on users who knowingly opt into using prerelease Rust
versions, and the only change is to allow this to work.

### Are there any considerations that need to be addressed for release?

I'm not aware of any.